Testing changes

+++ updateNear.sh
@@ -1,21 +1,24 @@
-##!/bin/bash
+#!/bin/bash
 
 source ~/.profile
+source $HOME/.cargo/env
 ip=$(curl ifconfig.me)
 network=$NEAR_NETWORK
 msg="msg"
 USER=$(whoami)
+nodekey=$(cat /home/$USER/near-ci/localnet/localnet-node0/validator_key.json | jq .public_key | tr -d '"')
 
 #check node version diff
 diff <(curl -s https://rpc.$network.near.org/status | jq .version) <(curl -s http://127.0.0.1:3030/status | jq .version)
 
 #start update if local version is different
-if [ $? -ne 0 ]; then
+if [ $? -eq 0 ]; then
     echo "start update";
     rustup default nightly
-    rm -rf /home/$USER/nearcore.new
+    [ -d /home/$USER/nearcore.new ] && rm -rf /home/$USER/nearcore.new
     version=$(curl -s https://rpc.$network.near.org/status | jq .version.version)
     strippedversion=$(echo "$version" | awk -F "\"" '{print $2}' | awk -F "-" '{print $1}')
+    mkdir /home/$USER/nearcore.new
     git clone --branch $strippedversion https://github.com/nearprotocol/nearcore.git /home/$USER/nearcore.new
     cd /home/$USER/nearcore.new
     make release
@@ -24,10 +27,10 @@ if [ $? -ne 0 ]; then
         if [ $? -eq 0 ]; then
         echo "new build was successfull, run test nodes"
 
-        /home/$USER/nearcore.new/target/release/neard --home /home/$USER/near-update/localnet/node0 run
-        /home/$USER/nearcore.new/target/release/neard --home /home/$USER/near-update/localnet/node1 run --boot-nodes ed25519:7PGseFbWxvYVgZ89K1uTJKYoKetWs7BJtbyXDzfbAcqX@127.0.0.1:24550
-        /home/$USER/nearcore.new/target/release/neard --home /home/$USER/near-update/localnet/node2 run --boot-nodes ed25519:7PGseFbWxvYVgZ89K1uTJKYoKetWs7BJtbyXDzfbAcqX@127.0.0.1:24550
-        /home/$USER/nearcore.new/target/release/neard --home /home/$USER/near-update/localnet/node3 run --boot-nodes ed25519:7PGseFbWxvYVgZ89K1uTJKYoKetWs7BJtbyXDzfbAcqX@127.0.0.1:24550
+        /home/$USER/nearcore.new/target/release/neard --home /home/$USER/near-ci/localnet/localnet-node0 run
+        /home/$USER/nearcore.new/target/release/neard --home /home/$USER/near-ci/localnet/localnet-node1 run --boot-nodes $nodekey@127.0.0.1:24550
+        /home/$USER/nearcore.new/target/release/neard --home /home/$USER/near-ci/localnet/localnet-node2 run --boot-nodes $nodekey@127.0.0.1:24550
+        /home/$USER/nearcore.new/target/release/neard --home /home/$USER/near-ci/localnet/localnet-node3 run --boot-nodes $nodekey@127.0.0.1:24550
         sleep 10
         echo "run test"        
         for count in {0..3}
@@ -42,20 +45,20 @@ if [ $? -ne 0 ]; then
                 ./twilio.sh "$msg"
 
                 #Remove testing data
-                rm -rf /home/$USER/near-update/localnet/node0/data
-                rm -rf /home/$USER/near-update/localnet/node1/data
-                rm -rf /home/$USER/near-update/localnet/node2/data
-                rm -rf /home/$USER/near-update/localnet/node3/data
+                rm -rf /home/$USER/near-ci/localnet/localnet-node0/data
+                rm -rf /home/$USER/near-ci/localnet/localnet-node1/data
+                rm -rf /home/$USER/near-ci/localnet/localnet-node2/data
+                rm -rf /home/$USER/near-ci/localnet/localnet-node3/data
                 exit 1
             fi
         done
         echo 'Testing localnet complete'
 
         #Remove testing data
-        rm -rf /home/$USER/near-update/localnet/node0/data
-        rm -rf /home/$USER/near-update/localnet/node1/data
-        rm -rf /home/$USER/near-update/localnet/node2/data
-        rm -rf /home/$USER/near-update/localnet/node3/data
+        rm -rf /home/$USER/near-ci/localnet/localnet-node0/data
+        rm -rf /home/$USER/near-ci/localnet/localnet-node1/data
+        rm -rf /home/$USER/near-ci/localnet/localnet-node2/data
+        rm -rf /home/$USER/near-ci/localnet/localnet-node3/data
         pkill neard
 
         echo 'test is succesfull, deploy a new node'
@@ -74,3 +77,5 @@ if [ $? -ne 0 ]; then
 
         fi
 fi
+
+
